@@ -1,5 +1,6 @@
 package com.openclassrooms.starterjwt.security.services;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -11,11 +12,13 @@ import com.openclassrooms.starterjwt.repository.UserRepository;
 
 @Service
 public class UserDetailsServiceImpl implements UserDetailsService {
+
+  @Autowired
   UserRepository userRepository;
 
-  UserDetailsServiceImpl(UserRepository userRepository) {
-    this.userRepository = userRepository;
-  }
+  //UserDetailsServiceImpl(UserRepository userRepository) {
+  //  this.userRepository = userRepository;
+  //}
 
   @Override
   @Transactional
